+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added saving software versions to a JSON file.
 - Added generating simulation id based on local date and time.
 - Added creating directory structure for simulation.
+- Added launching simulation.
 - Added hierarchy of SimTools exceptions.
 - Added package installer.
 - Added package description.